Updated Game's Score method to pass all unit tests.
Added functionality to check for spares and add the following rolls score as a bonus in the case of them.

Fixed a failing unit test in the check for Strikes which was counting scoring a 0 on the first roll and a 10 on the second roll as a strike which would calculate incorrectly, now makes sure it is the first throw of the round before counting as a strike.

Cleaned up and finalised code in the Game class.